Update README.md
Overview This project is aimed at detecting whether individuals in a vehicle are asleep or awake based on images or video input. The system is designed to identify multiple people in a single image or video frame and determine how many are sleeping. If a person is detected as sleeping, the system marks them in red and predicts their age. Additionally, a pop-up message is displayed, indicating the number of sleeping people and their ages.